+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed 
+
+* Comment or empty line between posting creating errors (#29) 
+
 
 ## [1.0.0-alpha.11] - 2023-06-15